tests/wpt/web-platform-tests/css/css-text/parsing/word-spacing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-spacing with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-spacing">
<meta name="assert" content="word-spacing supports the full grammar 'normal | <length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("word-spacing", "normal");

test_valid_value("word-spacing", "10px");
test_valid_value("word-spacing", "-20px");
test_valid_value("word-spacing", "calc(2em + 3ex)");
test_valid_value("word-spacing", "0", "0px");
</script>
</body>
</html>
